feat: move business logic to page insted of component
