+++ src/components/auth/sign-up-form/sign-up-form.tsx
@@ -1,61 +1,22 @@
 import { Box, Divider, Typography } from "@mui/material";
 import { CredentialResponse, GoogleLogin } from "@react-oauth/google";
-import { FetchBaseQueryError } from "@reduxjs/toolkit/query";
-import React, { useCallback, useEffect } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
-import { useNavigate } from "react-router-dom";
 
 import { BaseButton, CustomFormGroup } from "~/components/common/index.ts";
-import { AppRoute } from "~/libs/enum/app-route.enum.ts";
 import { useSignUpForm } from "~/libs/hooks/index.ts";
-import { useAddUserGoogleMutation } from "~/redux/auth/auth-api.ts";
-import { setUser } from "~/redux/auth/auth-slice.ts";
-import { useAppDispatch } from "~/redux/example-hooks.ts";
 import theme from "~/theme.ts";
 
 import { AuthLinks, SignInLink, StyledFormContainer } from "../index.ts";
 
-const SignUpForm: React.FC = () => {
+type Properties = {
+	onClick: (credentialResponse: CredentialResponse) => void;
+};
+
+const SignUpForm: React.FC<Properties> = ({ onClick }: Properties) => {
 	const { t } = useTranslation();
-	const dispatch = useAppDispatch();
-	const navigate = useNavigate();
-	const [addUser, { data, error, isError, isSuccess }] =
-		useAddUserGoogleMutation();
-	const {
-		control,
-		errors,
-		handleFormSubmit,
-		isLoading,
-		serverError,
-		setServerError,
-	} = useSignUpForm();
-
-	useEffect(() => {
-		if (isSuccess) {
-			dispatch(setUser(data));
-			navigate(AppRoute.VERIFY_EMAIL);
-		}
-		if (isError) {
-			const loadError = (error as FetchBaseQueryError).data
-				? ((error as FetchBaseQueryError).data as Error)
-				: { message: t("Error.unknowError") };
-			setServerError(loadError.message);
-		}
-	}, [data, dispatch, error, isError, isSuccess, navigate, setServerError, t]);
-
-	const onSuccess = useCallback(
-		async (credentialResponse: CredentialResponse) => {
-			try {
-				await addUser(credentialResponse);
-			} catch (error) {
-				const loadError = (error as FetchBaseQueryError).data
-					? ((error as FetchBaseQueryError).data as Error)
-					: { message: t("Error.unknowError") };
-				setServerError(loadError.message);
-			}
-		},
-		[addUser, setServerError, t],
-	);
+	const { control, errors, handleFormSubmit, isLoading, serverError } =
+		useSignUpForm();
 
 	return (
 		<>
@@ -66,11 +27,7 @@ const SignUpForm: React.FC = () => {
 				<Typography color={theme.palette.secondary.main} variant="dmSans">
 					{t("SignUpComponent.signUpWithGoogle")}
 				</Typography>
-				<GoogleLogin
-					logo_alignment="center"
-					onSuccess={onSuccess}
-					type="icon"
-				/>
+				<GoogleLogin logo_alignment="center" onSuccess={onClick} type="icon" />
 				<Divider sx={{ width: "100%" }}>
 					<Typography color="primary" variant="body1">
 						{t("SignUpComponent.or")}
